Only test passing None on 3.11+

+++ aiosqlite/tests/smoke.py
@@ -3,6 +3,7 @@
 
 import asyncio
 import sqlite3
+import sys
 from pathlib import Path
 from sqlite3 import OperationalError
 from tempfile import TemporaryDirectory
@@ -364,9 +365,10 @@ class SmokeTest(IsolatedAsyncioTestCase):
             with self.assertRaises(sqlite3.DatabaseError):
                 await db.execute("DROP TABLE test_drop")
 
-            # Disabling the authorizer re-enables DROP
-            await db.set_authorizer(None)
-            await db.execute("DROP TABLE test_drop")
+            if sys.version_info >= (3, 11):
+                # Disabling the authorizer re-enables DROP
+                await db.set_authorizer(None)
+                await db.execute("DROP TABLE test_drop")
 
     async def test_set_authorizer_exception_propagation(self):
         """Test that exceptions raised in authorizer callback are caught by SQLite"""
